+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## 0.0.1 (2023-06-18)
+
+* **semaphore:** add a semaphore implementation